pyproject.toml
[tool.poetry]
name = "transactions_processor"
version = "0.1.0"
description = ""
authors = ["Sharon Fabin <i.sharon.fabin@gmail.com>", "Ido Brosh <idobroshm@gmail.com>"]
packages = [{ include = "transactions_processor"}]
readme = "README.md"

[tool.poetry.scripts]
start = "transactions_processor.main:lambda_handler"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.26.4"
tabula-py = "^2.9.0"
openpyxl = "^3.1.2"
xlrd = "^2.0.1"
python-dotenv = "^1.0.1"
jpype1 = "^1.5.0"
aws-lambda-powertools = "^2.38.1"
python-multipart = "^0.0.9"
multipart = "^0.2.4"
boto3 = "^1.34.118"
requests-toolbelt = "^1.0.0"
pydantic = "^2.7.3"
pandas = "^2.2.2"
loguru = "^0.7.2"
beanie = "^1.27.0"
pydantic-settings = "^2.6.0"
aiohttp = "^3.10.10"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"
log_cli = true
log_cli_level = "CRITICAL"
log_cli_format = "%(message)s"

log_file = "tests/pytest.log"
log_file_level = "DEBUG"
log_file_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_file_date_format = "%Y-%m-%d %H:%M:%S"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
